:fire: First Sucess Third Rendering Method :fire:
Rendering consist of chunk-meshes of 16x16x16 blocks, all merged into
one geometry (with the same material).

This material is created using TextureStitcher's image, and each 3d face
has it's faceVertexUvs changed to set the texture (including ao) and
then merged into it's correspondent chunk.

Progress: Managed to get blocks to show up with specific texture, no
ambient occlusion, but it's already in the texture

Chunks are needed due to frustumCulling: if they weren't there to
separate the mesh we wouldn't be able to hide some blocks when the
camera isn't looking